sync

+++ deleteComment.php
@@ -3,7 +3,7 @@
   $data = json_decode(file_get_contents('php://input'));
   $userName = mysqli_real_escape_string($link, $data->{'userName'});
   $passhash = mysqli_real_escape_string($link, $data->{'passhash'});
-  $commentID = (int)mysqli_real_escape_string($link, $data->{'commentID'});
+  $commentID = mysqli_real_escape_string($link, $data->{'commentID'});
   $sql = 'SELECT * FROM users WHERE name LIKE "'.$userName.'" AND passhash = "'.$passhash.'";';
   $res = mysqli_query($link, $sql);
   $success = false;
