+++ incrementViews.php
@@ -1,7 +1,7 @@
 <?
   require("db.php");
   $data = json_decode(file_get_contents('php://input'));
-  $trackID = (int)mysqli_real_escape_string($link, $data->{'trackID'});
+  $trackID = mysqli_real_escape_string($link, $data->{'trackID'});
 
   function ipToDec($ip){
     $parts=explode(".",$ip);
